Refactored TopRatedMoviesFragment to use ViewModel
# created BaseMoviesFragment.kt
# used a single repo and different instances instantiated, hold and provided by Hilt
# crated annotations do sign which instances Hilt should inject
# created base ViewModel
