+++ app/src/main/java/com/curtesmalteser/popularmoviesstage1/repository/IMoviesProvider.kt
@@ -18,3 +18,11 @@ class PopularMoviesProvider @Inject constructor(private val moviesAPI: MoviesAPI
             moviesAPI.fetchPopularMovies(queryParams)
         }
 }
+
+class TopRatedMoviesProvider @Inject constructor(private val moviesAPI: MoviesAPIInterface) :
+    IMoviesProvider {
+    override suspend fun fetchMovies(queryParams: Map<String, String>): Result<MoviesModel> =
+        runCatching {
+            moviesAPI.fetchTopRated(queryParams)
+        }
+}
